Database creation complete
We can now have a different module taking care of updating the db
It's not yet connected to the front end, but it's in place.
Because of the the `if __name__` check, we can run this module
from the root directory of the project (next to run_server) as
`python -m my_app.get_json` to see it in action. It takes some time,
and creates a db.json tinyDB database in the root of the directory.

I have now hardcoded "hot" sort_parameter in the url of the subreddit.

database.py was not needed as settings.py is already there