Update Working architecture, faster results
This update now works according to the new posted architecture as
described in #67. Just make sure that you populate the database by
running `python -m music_app.get_json` which will create the
database file.

api_access is still there in the directory, but it is not required
anymore.

The get_songs implementation that I have wrote in views.py, currently
only gets the results as described by the search term, the sort and
search type are disabled at the moment. We need to implement it.