Run get_json with an interval periodically in a thread
This is pretty complicated, and we *MUST* move to celery as
Tyler mentioned. Some things that are to be noted is that
you will see the logger messages printed two times, as now
there are two instances of logger, ie the thread and the main
process, I don't think the code is running two times.

To kill the server and restart, Ctrl-C will no longer work, as there
are threads now. Do the following:

1) Find the process id of the task by running
`ps -ef | grep run_server`
2) Kill all the process using `kill -9 PID` where PID is the number
you found in 1.

Also auto_restart of flask will not work the way it should. Sorry
about it. Manually restart with the process mentioned above.