Changing the interface of get_json
Previosly `get_json_from_subreddit` took `sort_type` as the only
parameter, which was changed in the last commit. However I realized
that the url is always the same, however, the sort_type and sometimes
the limit parameter will be changing. So it has been added as a
parameter.

The main code at the bottom is also changed to reflect this fact.