Fixtures related testing.
Now we can mock the tests cases, to download a fixtured version
of json, which means, the requests library will not actually go
to the website but actually fetch the json from `tests/fixtures/`
directory and send that as a response.

It uses the `responses` library, which has been added to the
`dev-requirements` to automatically download it using pip.

`tests/__init__.py` is changed for setting up the fixtures,
and now get_json is tested with the new fixtured responses.